# Speed up TimeSeriesEvaluator